feat(ui): add UserAvatar primitive per D2 and adopt it at the 3 ad-hoc call sites
OrganizationAvatar (org logo + initials fallback) has been the one true
avatar component; person avatars were ad-hoc <img src={avatarUrl}> with no
initials fallback. Adds UserAvatar to ui/src/primitives/ (not literally
alongside OrganizationAvatar, which lives in product-client — product-ui
also needs to render it, and the dependency direction only allows
product-client -> product-ui -> ui, not the reverse), mirroring
OrganizationAvatar's className-driven sizing convention and absorbing the
avatarFailed-on-error fallback logic that AccountSettingsPane hand-rolled.

Of the 3 call sites named in the catalog, ProductSidebarAccountFooter.tsx
(product-ui) no longer exists — it was dead code deleted in this same cull —
so only 2 real call sites get rewired: SidebarAccountFooter.tsx
(product-client) and AccountSettingsPane.tsx's AccountAvatar/initialsForName
helpers (product-ui), both now composing UserAvatar directly.
